% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnover.R
\name{detrend_chordia_2007}
\alias{detrend_chordia_2007}
\title{Detrend time-series as outlined in Chordia et al. (2007)}
\usage{
detrend_chordia_2007(turnover, month, t)
}
\description{
Detrend time-series as outlined in Chordia et al. (2007)
}
