% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends_download.R
\name{int_split}
\alias{int_split}
\title{Split a time interval in chunks of same duration}
\usage{
int_split(int, chunk_duration = lubridate::days(180))
}
\arguments{
\item{int}{An interval object}

\item{chunk_duration}{Duration of interval}
}
\description{
Split a time interval in chunks of same duration
}
